commit2

+++ src/models.py
@@ -12,7 +12,7 @@ class Person(Base):
     __tablename__ = 'user'
     # Here we define columns for the table person
     # Notice that each column is also a normal Python instance attribute.
-    id = Column(Integer, primary_key=True)
+    id = Column(Integer , primary_key=True)
     name = Column(String(250), nullable=False)
     user_name = Column(String(250), nullable=False)
     email = Column(String(50), nullable=False)
@@ -26,11 +26,11 @@ class Character(Base):
     # Notice that each column is also a normal Python instance attribute.
     id = Column(Integer, primary_key=True)
     name = Column(String(250))
-    height = Column(Integer(250))
+    height = Column(Integer)
     hair_color = Column(String(10), nullable=False)
     skin_color = Column(String(10))
     eye_color = Column(String(50))
-    birth_year = Column(Integer(8))
+    birth_year = Column(Integer)
     gender = Column(String(10))
 
 
@@ -44,14 +44,14 @@ class Planets(Base):
     # Notice that each column is also a normal Python instance attribute.
     id = Column(Integer, primary_key=True)
     name = Column(String(50))
-    rotation_period = Column(Integer(10))
-    orbital_period = Column(Integer(10))
-    diameter = Column(Integer(10))
+    rotation_period = Column(Integer)
+    orbital_period = Column(Integer)
+    diameter = Column(Integer)
     climate = Column(String(20))
     gravity = Column(String(30))
     terrain = Column(String(30))
-    surface_water = Column(Integer(5))
-    population = Column(Integer(100))
+    surface_water = Column(Integer)
+    population = Column(Integer)
     def to_dict(self):
         return {}
 
@@ -63,11 +63,11 @@ class Species(Base):
     name = Column(String(50))
     classification = Column(String(50))
     designation = Column(String(50))
-    average_height = Column(Integer(50))
+    average_height = Column(Integer)
     skin_colors = Column(String(200))
     hair_colors = Column(String(200))
-    average_lifespan = Column(Integer(50))
-    language = Column(Integer(80))
+    average_lifespan = Column(Integer)
+    language = Column(Integer)
     def to_dict(self):
         return {}
 
@@ -77,7 +77,6 @@ class Favorite(Base):
     user_id = Column(Integer, ForeignKey('user.id'))
     character_id = Column(String(250), ForeignKey('character.id'))
     planet_id = Column((String(250)), ForeignKey('planet.id'))
-
 
     def to_dict(self):
         return {}
